+++ src/systems.py
@@ -18,7 +18,7 @@ def _entity_req( self, req ):
 
 
 def label( self, glbl ):
-	req = ["pos", "size", "text", "font", "font_size", "font_color", "font_align", "text_align"]
+	req = ["pos", "rect", "text", "font", "font_size", "font_color", "font_align", "text_align"]
 	e_main, e_ind = _entity_req( self, req )
 	for dt in e_main:
 		font = p.font.SysFont( dt["font"], dt["font_size"] )
@@ -26,24 +26,24 @@ def label( self, glbl ):
 		dimn = font.size( dt["text"] )
 		pkx = {  # Position Key X
 			"left"  :dt["pos"][0],
-			"center":dt["pos"][0]+dt["size"][0]/2-(dimn[0]/2 if dt["text_align"] else 0),
-			"right" :dt["pos"][0]+dt["size"][0]-(dimn[0] if dt["text_align"] else 0),
+			"center":dt["pos"][0]+dt["rect"][0]/2-(dimn[0]/2 if dt["text_align"] else 0),
+			"right" :dt["pos"][0]+dt["rect"][0]-(dimn[0] if dt["text_align"] else 0),
 		}
 		pky = {  # Position Key Y
 			"top"   :dt["pos"][1],
-			"center":dt["pos"][1]+dt["size"][1]/2-(dimn[1]/2 if dt["text_align"] else 0),
-			"bottom":dt["pos"][1]+dt["size"][1]-(dimn[1] if dt["text_align"] else 0),
+			"center":dt["pos"][1]+dt["rect"][1]/2-(dimn[1]/2 if dt["text_align"] else 0),
+			"bottom":dt["pos"][1]+dt["rect"][1]-(dimn[1] if dt["text_align"] else 0),
 		}
 		glbl["surf"].blit( label, (pkx[dt["font_align"]["x"]], pky[dt["font_align"]["y"]]) )
 
 def rect( self, glbl ):
-	req = ["pos", "size", "color", ]
+	req = ["pos", "rect", "color", ]
 	e_main, e_ind = _entity_req( self, req )
 	for dt in e_main:
-		p.draw.rect(glbl["surf"], dt["color"], [dt["pos"][0], dt["pos"][1], dt["size"][0], dt["size"][1]])
+		p.draw.rect(glbl["surf"], dt["color"], [dt["pos"][0], dt["pos"][1], dt["rect"][0], dt["rect"][1]])
 
 def click( self, glbl ):
-	req = ["pos", "size", "clicked", ]
+	req = ["pos", "rect", "clicked", ]
 	e_main, e_ind = _entity_req( self, req )
 
 	clicked = False
@@ -52,8 +52,8 @@ def click( self, glbl ):
 		if c["clicked"]: clicked = True
 
 	for eid, dt in zip(reversed(self.entity( self.ENTITIES, req)), reversed(e_main)):
-		if dt["pos"][0] < glbl["event"]["mPos"][0] < dt["pos"][0]+dt["size"][0] and not clicked and \
-				dt["pos"][1] < glbl["event"]["mPos"][1] < dt["pos"][1]+dt["size"][1] and glbl["event"]["mTrgOn"]:  # main
+		if dt["pos"][0] < glbl["event"]["mPos"][0] < dt["pos"][0]+dt["rect"][0] and not clicked and \
+				dt["pos"][1] < glbl["event"]["mPos"][1] < dt["pos"][1]+dt["rect"][1] and glbl["event"]["mTrgOn"]:  # main
 			dt["clicked"] = True
 			clicked = True
 		if glbl["event"]["mTrgOff"] and dt["clicked"]: dt["clicked"] = False
@@ -72,7 +72,7 @@ def move( self, glbl ):
 				dt["placement_ofs"] = [None, None]
 		self.entity_save( eid, dt )
 
-def move_child( self, glbl ):
+def moveChild( self, glbl ):
 	req = ["pos", "movable", "child", "placement_ofs"]
 	e_main, e_ind = _entity_req( self, req )
 	for eid, dt in zip( reversed( e_ind ), reversed( e_main ) ):
@@ -88,7 +88,7 @@ def move_child( self, glbl ):
 
 # render isolated wire (wire connected to the mouse)
 def connectorWireIso( self, glbl ):
-	req = ["pos", "size", "clicked", "connectee" ]
+	req = ["pos", "rect", "clicked", "connectee" ]
 	e_main, e_ind = _entity_req( self, req )
 	for dt in e_main:
 		if dt["clicked"]:
@@ -96,7 +96,7 @@ def connectorWireIso( self, glbl ):
 
 # render connected wire (wire connected to the connected node)
 def connectorWireMrg( self, glbl ):
-	req = [ "pos", "size", "connectee" ]
+	req = [ "pos", "rect", "connectee" ]
 	e_main, e_ind = _entity_req( self, req )
 	for dt in e_main:
 		for cncte in dt["connectee"]:
@@ -105,7 +105,7 @@ def connectorWireMrg( self, glbl ):
 
 # connects each nodes
 def connectNode( self, glbl ):
-	req = ["obj_id", "pos", "size", "clicked", "connectee", "connect_en", "connect_tg" ]
+	req = ["obj_id", "pos", "rect", "clicked", "connectee", "connect_en", "connect_tg" ]
 	e_main, e_ind = _entity_req( self, req )
 
 	connector, connectee, cmpct_id = None, None, None
@@ -113,8 +113,8 @@ def connectNode( self, glbl ):
 		if dt["clicked"]:  # get the entity connector
 			connector = (eid, dt)
 
-		if dt["pos"][0] < glbl["event"]["mPos"][0] < dt["pos"][0]+dt["size"][0] and \
-				dt["pos"][1] < glbl["event"]["mPos"][1] < dt["pos"][1]+dt["size"][1] and glbl["event"]["mTrgOff"]:
+		if dt["pos"][0] < glbl["event"]["mPos"][0] < dt["pos"][0]+dt["rect"][0] and \
+				dt["pos"][1] < glbl["event"]["mPos"][1] < dt["pos"][1]+dt["rect"][1] and glbl["event"]["mTrgOff"]:
 			connectee = (eid, dt)
 			cmpct_id = self.TAG[ self.ENTITIES.index( eid ) ] # compact entity id for connector
 
@@ -124,15 +124,90 @@ def connectNode( self, glbl ):
 			connector[1]["connectee"].append(cmpct_id)
 			self.entity_save( connector[0], connector[1])
 
+# render editable input field
+def shwCursor( self, glbl ):  # TODO
+	req = ["pos", "rect", "text", "font", "font_size", "at", "cursor", "font_align"]
+	e_main, e_ind = _entity_req( self, req )
+
+	for eid, dt in zip( reversed( e_ind ), reversed( e_main ) ):
+		if dt["at"]:
+			font = p.font.SysFont( dt["font"], dt["font_size"] )
+			dimn = font.size( dt["text"] )
+
+			pkx = {  # Position Key X
+				"left"  :dt["pos"][0],
+				"center":dt["pos"][0]+dt["rect"][0]/2-(dimn[0]/2 if dt["text_align"] else 0),
+				"right" :dt["pos"][0]+dt["rect"][0]-(dimn[0] if dt["text_align"] else 0),
+			}
+			pky = {  # Position Key Y
+				"top"   :dt["pos"][1],
+				"center":dt["pos"][1]+dt["rect"][1]/2-(dimn[1]/2 if dt["text_align"] else 0),
+				"bottom":dt["pos"][1]+dt["rect"][1]-(dimn[1] if dt["text_align"] else 0),
+			}
+
+			if glbl["chrc"]["raw"] == p.K_RIGHT and dt["cursor"] < len(dt["text"]): dt["cursor"] += 1
+			elif glbl["chrc"]["raw"] == p.K_LEFT and dt["cursor"] > 0: dt["cursor"] -= 1
+
+			gcur = glbl["cursor"]
+
+			if glbl["cursor"]["blink"]:
+				txt = font.render( dt["text"][0:dt["cursor"]], False, (0, 0, 0) )
+				textRect = txt.get_rect()
+				locx = textRect[2]+gcur["ofsx"]+pkx[dt["font_align"]["x"]]
+				p.draw.rect( glbl["surf"], (0, 0, 0), (locx, pky[dt["font_align"]["y"]], gcur["width"], dt["font_size"]) )
+			self.entity_save( eid, dt )
+
+# edit the /text/ component
+def editText( self, glbl ):  # TODO
+	req = ["text", "at", "cursor"]
+	e_main, e_ind = _entity_req( self, req )
+
+	for eid, dt in zip( reversed( e_ind ), reversed( e_main ) ):
+		if dt["at"]:
+			if glbl["chrc"]["unicode"] in glbl["chrc"]["valid"] and glbl["chrc"]["unicode"] != "":
+				l = list(dt["text"])
+				l.insert( dt["cursor"], glbl["chrc"]["unicode"] )
+				dt["text"] = "".join(l)
+				dt["cursor"] += 1
+			if glbl["chrc"]["unicode"] == chr(p.K_BACKSPACE) and dt["cursor"] > 0:
+				l = list( dt["text"] )
+				del l[dt["cursor"]]
+				dt["text"] = "".join( l )
+				dt["cursor"] -= 1
+			self.entity_save( eid, dt )
+
+# when user click ONCE on the object, this will always be true unless the user click away the object's dimension
+def at( self, glbl ):
+	req = ["pos", "rect", "at", ]
+	e_main, e_ind = _entity_req( self, req )
+
+	at = False
+	e_click = _entity_only( self, ["at"] ) # check if any other entity has clicked
+	for c in e_click:
+		if c["at"]: at = True
+
+	for eid, dt in zip(reversed(self.entity( self.ENTITIES, req)), reversed(e_main)):
+		if dt["pos"][0] < glbl["event"]["mPos"][0] < dt["pos"][0]+dt["rect"][0] and not at and\
+				dt["pos"][1] < glbl["event"]["mPos"][1] < dt["pos"][1]+dt["rect"][1]:  # if cursor is at the object
+			if glbl["event"]["mTrgOn"]:  # cursor on clicked
+				dt["at"] = True
+				glbl["cursor"]["txt_inp"] = True
+				self.entity_save( eid, dt )
+				break
+			# else: cursor hovering over the object or off clicked
+		else:  # not at the object
+			if glbl["event"]["mTrgOn"]:  # cursor on clicked
+				dt["at"] = False
+				glbl["cursor"]["txt_inp"] = False
+				self.entity_save( eid, dt )
+
 
 # special systems -- for specific purposes
 
-def gen_fields( self, glbl ):
-	req = ["obj_id", "pos", "field", "child" ]
+def genFields( self, glbl ):
+	req = ["obj_id", "pos", "rect", "field", "child" ]
 	e_main, e_ind = _entity_req( self, req )
 	for dt in e_main:
-		# TODO
-		# ent = self.entity_equal( self.entity( self.ENTITIES, "fld_nm" ), obj_id=dt["child"] )  # get entity fields that has the same obj_id as the child's
 		e_missing = []  # field names that are missing
 
 		entc_req = self.entity_pp_strip( self.entity_data(self.ENTITIES, req=["obj_id", "fld_nm"]), strict=False, obj_id=dt["child"] )
@@ -149,15 +224,15 @@ def gen_fields( self, glbl ):
 		ind_i, ind_o = 0, 0
 		for f in e_missing:  # creating a new entity
 			if dt["field"][f] == glbl["node_typ"]["inp"]:
-				prfb.BoxInpField( self, dt["child"], dt["pos"], dt["size"], nd_pd, nd_sz, ind_i,
+				prfb.BoxInpField( self, dt["child"], dt["pos"], dt["rect"], nd_pd, nd_sz, ind_i,
 				                  dt["field"][f], f, [prfb.NdEn.BoxOut] )  # going for reciprocal b/c inp cnnct to out
 				ind_i += 1
 			elif dt["field"][f] == glbl["node_typ"]["out"]:
-				prfb.BoxOutField( self, dt["child"], dt["pos"], dt["size"], nd_pd, nd_sz, ind_o,
+				prfb.BoxOutField( self, dt["child"], dt["pos"], dt["rect"], nd_pd, nd_sz, ind_o,
 				                  dt["field"][f], f, [prfb.NdEn.BoxInp] )
 				ind_o += 1
 			elif dt["field"][f] == glbl["node_typ"]["user"]:  # user defined node, uses custom entities
 				ind_u = max( ind_i, ind_o )+1
-				self.create(obj_id=dt["child"], pos=[dt["pos"][0], dt["pos"][1]+ind_u*(nd_sz+nd_pd)+nd_pd+dt["size"][1]/4+nd_pd],
-				            size=[dt["size"][0], 25], color=(0, 0, 0), fld_nm=f, fld_typ=dt["field"][f], fld_dt=None )
+				self.create(obj_id=dt["child"], pos=[dt["pos"][0], dt["pos"][1]+ind_u*(nd_sz+nd_pd)+nd_pd+dt["rect"][1]/4+nd_pd],
+				            rect=[dt["rect"][0], 25], color=(0, 0, 0), fld_nm=f, fld_typ=dt["field"][f], fld_dt=None )